+++ examples/2-template_switch/Controller.php
@@ -1,0 +1,16 @@
+<?php
+
+use App\Http\Controllers\Controller as BaseController;
+use ArrowsStepper;
+
+class Controller extends BaseController
+{
+    public function index(ArrowsStepper $stepper)
+    {
+        $stepper->setCurrentStepName('step 3');
+
+        return $this->view('example', [
+            'stepper' => $stepper->render()
+        ]);
+    }
+}